pyproject.toml
[tool.poetry]
name = "meteval"
version = "1.0.0"
description = "Image analyser and validator microservice component"
authors = ["Karel Vlk <kaja.vlk@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[[tool.poetry.source]]
name = "torchcpu"
url = "https://download.pytorch.org/whl/cpu/"
priority = "supplemental"

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["api"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12.0"
scipy = "^1.10.1"
scikit-learn = "1.3.2"
pillow = "^9.5.0"
opencv-python = "^4.7.0.72"
pandas = "^2.0.2"
fastapi = {version="^0.100.0", extras = ["all"]}
routes = "^2.5.1"
numpy = "1.23.2"
uvicorn = {version = "^0.22.0", extras = ["standard"]}
python-multipart = "^0.0.6"
requests-toolbelt = "^1.0.0"
scikit-image = "^0.21.0"
matplotlib = "^3.7.1"
tqdm = "^4.65.0"
parse = "^1.19.1"
ultralytics = "8.0.123"
torchaudio = "^2.0.2"
debugpy = "1.6.7"
orjson = "^3.9.5"
pebble = "^5.0.3"
albumentations = "^1.3.1"
httpx = "^0.26.0"
toml = "^0.10.2"
tritonclient = { version="2.42.0", extras = ["grpc"] }
torchvision = {version = "^0.15.2", source = "torchcpu"}
torch = {version = "2.0.1", source = "torchcpu"}


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ruff = "0.*"
mypy = "^1.3.0"
black = "^23.3.0"
pylint = "^2.17.4"
pytest-cov = "^4.1.0"
pytest = "^7.3.2"
jsonschema = "^4.17.3"
pre-commit = "^3.3.3"
pydantic = "^2.1.1"


[tool.ruff]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "B011",  # use AssertionError instead of assert
]
src = ["src", "tests"]
target-version = "py311"
